front/back: fixed dates issue and proper form-filling

+++ server/database/migrations/2025_08_06_142628_add_renter_addresses_to_renters_table.php
@@ -12,9 +12,17 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('renters', function (Blueprint $table) {
-            $table->string('id_card_address')->nullable();
-            $table->string('driver_license_address')->nullable();
-            $table->string('passport_address')->nullable();
+            if (!Schema::hasColumn('renters', 'id_card_address')) {
+                $table->string('id_card_address')->nullable();
+            }
+        
+            if (!Schema::hasColumn('renters', 'driver_license_address')) {
+                $table->string('driver_license_address')->nullable();
+            }
+        
+            if (!Schema::hasColumn('renters', 'passport_address')) {
+                $table->string('passport_address')->nullable();
+            }
         });
     }
 
@@ -24,9 +32,15 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('renters', function (Blueprint $table) {
-            $table->dropColumn('id_card_address');
-            $table->dropColumn('driver_license_address');
-            $table->dropColumn('passport_address');
+            if (Schema::hasColumn('renters', 'id_card_address')) {
+                $table->dropColumn('id_card_address');
+            }
+            if (Schema::hasColumn('renters', 'driver_license_address')) {
+                $table->dropColumn('driver_license_address');
+            }
+            if (Schema::hasColumn('renters', 'passport_address')) {
+                $table->dropColumn('passport_address');
+            }
         });
     }
 };
